requirement.txt
bcrypt==3.2.0    
Flask==1.1.2
Flask-Assets==2.0
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.8.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.4.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
SQLAlchemy==1.3.21
pyjwt==1.7.1
pyhibp
cryptography